[UPDATE] chat-room-delete-authority 방장인 경우에만 방 삭제 가능하도록 구현

+++ chat/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render, redirect
 from .models import Message
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
+
 
 def index(request):
     messages = Message.objects.all()
@@ -22,6 +25,14 @@ def room(request, room_name):
 
 
 
+@login_required
 def room_delete(request, room_name):
+    # 방장 여부 확인
+    room_owner = Message.objects.filter(room=room_name, is_owner=True, user=request.user).first()
+    if not room_owner:
+        return HttpResponseForbidden("당신은 방장이 아닙니다.")
+
+    # 방장인 경우 방의 모든 메시지 삭제
     Message.objects.filter(room=room_name).delete()
+
     return redirect('chat:index')
